Add paddlespeech.cls and esc50 example.
