--- audio/.style.yapf
@@ -1,3 +1,0 @@
-[style]
-based_on_style = pep8
-column_limit = 80
