+++ paddlespeech/cls/datasets/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
 #
-# Licensed under the Apache License, Version 2.0 (the "License");
+# Licensed under the Apache License, Version 2.0 (the "License"
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
@@ -11,5 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .backends import *
-from .features import *
+from .esc50 import ESC50
+from .gtzan import GTZAN
+from .tess import TESS
+from .urban_sound import UrbanSound8K
+
+__all__ = [
+    'ESC50',
+    'UrbanSound8K',
+    'GTZAN',
+    'TESS',
+]
